refactor: keep Task & TaskType in the same module, add pydoc
