+++ application/backend/app/models/__init__.py
@@ -17,7 +17,7 @@ from .dataset_item import (
 from .label import Label, LabelReference
 from .model_revision import ModelRevision, TrainingInfo, TrainingStatus
 from .pipeline import Pipeline, PipelineStatus
-from .project import Project, Task
+from .project import Project
 from .shape import FullImage, Point, Polygon, Rectangle, Shape
 from .sink import (
     DisconnectedSinkConfig,
@@ -40,7 +40,7 @@ from .source import (
     VideoFileSourceConfig,
     WebcamSourceConfig,
 )
-from .task_type import TaskType
+from .task import Task, TaskType
 
 __all__ = [
     "ConfidenceThresholdDataCollectionPolicy",
